Use management command for database init
