+++ webapp/apps/billing/tests/test_models.py
@@ -5,11 +5,10 @@ from django.contrib.auth import get_user_model
 
 import pytest
 
-from webapp.apps.billing.models import (get_billing_data,
-                                        Customer, Plan, Product,
+from webapp.apps.billing.models import (Customer, Plan, Product,
                                         Subscription, SubscriptionItem,
                                         UsageRecord)
-
+from webapp.apps.billing.utils import get_billing_data
 
 User = get_user_model()
